requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
exceptiongroup==1.2.0
fastapi==0.109.0
graphql-core==3.2.3
greenlet==3.0.3
h11==0.14.0
httpie==3.2.2
idna==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.4
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlmodel==0.0.14
starlette==0.35.1
strawberry-graphql==0.219.1
typing_extensions==4.9.0
urllib3==2.2.0
uvicorn==0.27.0.post1
